Made engine.cpp compilable. Most of the work is in the physics engine.

+++ missile.h
@@ -1,17 +1,18 @@
 #ifndef __MISSILE_H__
 #define __MISSILE_H__
+#include "coretypes.h"
 #include "entity.h"
 
 class Missile: public Entity
 {
-	virtual Vector2f getSize() const;
+	virtual Vector2d getSize() const;
 	virtual void draw(sf::RenderTarget target) const;
-	virtual Vector2f movement(Int64 tm);
+	virtual Vector2d movement(Int64 tm);
 	virtual void event_received(EngineEvent *event);
 
 	private:
 	static const float speed = 3;
-	float x, y, angle;
+	double x, y, angle;
 	int playerUID;
 };
 #endif
